Fix (Step2): Lotto 생성 로직 LottoShop 에 포함 && 발권 개수 검증 로직 분리
